Move ex-overlay component out of overlay folder
